feat: allow podman to be used by just if available
Originally I wanted to implement the check with `command -v podman`, but
I was unsuccessful in evaluating the variable this way because the exit
status would have been 1 if the command would have been missing and just
errors out in this case.
I am unsure if `whereis` works everywhere, but `which` suffers from the
same.